+++ db-practice/src/main/java/com/si/practice/dbpractice/app/t.java
@@ -1,0 +1,4 @@
+package com.si.practice.dbpractice.app;
+
+public class t {
+}
