+++ db-practice/src/main/java/com/si/practice/dbpractice/client/order/api/OrderCreateService.java
@@ -1,0 +1,13 @@
+package com.si.practice.dbpractice.client.order.api;
+
+
+import com.si.practice.dbpractice.common.order.OrderDTO;
+
+import java.util.List;
+
+public interface OrderCreateService {
+
+    Long createOrder(OrderDTO orderDTO);
+
+    Boolean batchInsert(List<OrderDTO> orderDTOS);
+}
