+++ db-practice/src/main/java/com/si/practice/dbpractice/client/order/api/OrderUpdateService.java
@@ -1,0 +1,6 @@
+package com.si.practice.dbpractice.client.order.api;
+
+public interface OrderUpdateService {
+
+    Boolean updateOrderById(Long orderId, Integer status);
+}
