+++ db-practice/src/main/java/com/si/practice/dbpractice/client/order/request/QueryOrderRequest.java
@@ -1,0 +1,6 @@
+package com.si.practice.dbpractice.client.order.request;
+
+import java.io.Serializable;
+
+public class QueryOrderRequest implements Serializable {
+}
