+++ db-practice/src/main/java/com/si/practice/dbpractice/infrature/mysql/entity/DemandOrderExample.java
@@ -1,0 +1,858 @@
+package com.si.practice.dbpractice.infrature.mysql.entity;
+
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+
+public class DemandOrderExample {
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    protected String orderByClause;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    protected boolean distinct;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    protected List<Criteria> oredCriteria;
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public DemandOrderExample() {
+        oredCriteria = new ArrayList<Criteria>();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public void setOrderByClause(String orderByClause) {
+        this.orderByClause = orderByClause;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public String getOrderByClause() {
+        return orderByClause;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public void setDistinct(boolean distinct) {
+        this.distinct = distinct;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public boolean isDistinct() {
+        return distinct;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public List<Criteria> getOredCriteria() {
+        return oredCriteria;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public void or(Criteria criteria) {
+        oredCriteria.add(criteria);
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public Criteria or() {
+        Criteria criteria = createCriteriaInternal();
+        oredCriteria.add(criteria);
+        return criteria;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public Criteria createCriteria() {
+        Criteria criteria = createCriteriaInternal();
+        if (oredCriteria.size() == 0) {
+            oredCriteria.add(criteria);
+        }
+        return criteria;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    protected Criteria createCriteriaInternal() {
+        Criteria criteria = new Criteria();
+        return criteria;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public void clear() {
+        oredCriteria.clear();
+        orderByClause = null;
+        distinct = false;
+    }
+
+    /**
+     * This class was generated by MyBatis Generator.
+     * This class corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    protected abstract static class GeneratedCriteria {
+        protected List<Criterion> criteria;
+
+        protected GeneratedCriteria() {
+            super();
+            criteria = new ArrayList<Criterion>();
+        }
+
+        public boolean isValid() {
+            return criteria.size() > 0;
+        }
+
+        public List<Criterion> getAllCriteria() {
+            return criteria;
+        }
+
+        public List<Criterion> getCriteria() {
+            return criteria;
+        }
+
+        protected void addCriterion(String condition) {
+            if (condition == null) {
+                throw new RuntimeException("Value for condition cannot be null");
+            }
+            criteria.add(new Criterion(condition));
+        }
+
+        protected void addCriterion(String condition, Object value, String property) {
+            if (value == null) {
+                throw new RuntimeException("Value for " + property + " cannot be null");
+            }
+            criteria.add(new Criterion(condition, value));
+        }
+
+        protected void addCriterion(String condition, Object value1, Object value2, String property) {
+            if (value1 == null || value2 == null) {
+                throw new RuntimeException("Between values for " + property + " cannot be null");
+            }
+            criteria.add(new Criterion(condition, value1, value2));
+        }
+
+        public Criteria andIdIsNull() {
+            addCriterion("`id` is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdIsNotNull() {
+            addCriterion("`id` is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdEqualTo(Long value) {
+            addCriterion("`id` =", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdNotEqualTo(Long value) {
+            addCriterion("`id` <>", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdGreaterThan(Long value) {
+            addCriterion("`id` >", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdGreaterThanOrEqualTo(Long value) {
+            addCriterion("`id` >=", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdLessThan(Long value) {
+            addCriterion("`id` <", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdLessThanOrEqualTo(Long value) {
+            addCriterion("`id` <=", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdIn(List<Long> values) {
+            addCriterion("`id` in", values, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdNotIn(List<Long> values) {
+            addCriterion("`id` not in", values, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdBetween(Long value1, Long value2) {
+            addCriterion("`id` between", value1, value2, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdNotBetween(Long value1, Long value2) {
+            addCriterion("`id` not between", value1, value2, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerIsNull() {
+            addCriterion("`passenger` is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerIsNotNull() {
+            addCriterion("`passenger` is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerEqualTo(String value) {
+            addCriterion("`passenger` =", value, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerNotEqualTo(String value) {
+            addCriterion("`passenger` <>", value, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerGreaterThan(String value) {
+            addCriterion("`passenger` >", value, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerGreaterThanOrEqualTo(String value) {
+            addCriterion("`passenger` >=", value, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerLessThan(String value) {
+            addCriterion("`passenger` <", value, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerLessThanOrEqualTo(String value) {
+            addCriterion("`passenger` <=", value, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerLike(String value) {
+            addCriterion("`passenger` like", value, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerNotLike(String value) {
+            addCriterion("`passenger` not like", value, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerIn(List<String> values) {
+            addCriterion("`passenger` in", values, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerNotIn(List<String> values) {
+            addCriterion("`passenger` not in", values, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerBetween(String value1, String value2) {
+            addCriterion("`passenger` between", value1, value2, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerNotBetween(String value1, String value2) {
+            addCriterion("`passenger` not between", value1, value2, "passenger");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdIsNull() {
+            addCriterion("`prod_id` is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdIsNotNull() {
+            addCriterion("`prod_id` is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdEqualTo(Long value) {
+            addCriterion("`prod_id` =", value, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdNotEqualTo(Long value) {
+            addCriterion("`prod_id` <>", value, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdGreaterThan(Long value) {
+            addCriterion("`prod_id` >", value, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdGreaterThanOrEqualTo(Long value) {
+            addCriterion("`prod_id` >=", value, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdLessThan(Long value) {
+            addCriterion("`prod_id` <", value, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdLessThanOrEqualTo(Long value) {
+            addCriterion("`prod_id` <=", value, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdIn(List<Long> values) {
+            addCriterion("`prod_id` in", values, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdNotIn(List<Long> values) {
+            addCriterion("`prod_id` not in", values, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdBetween(Long value1, Long value2) {
+            addCriterion("`prod_id` between", value1, value2, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andProdIdNotBetween(Long value1, Long value2) {
+            addCriterion("`prod_id` not between", value1, value2, "prodId");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusIsNull() {
+            addCriterion("`status` is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusIsNotNull() {
+            addCriterion("`status` is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusEqualTo(Integer value) {
+            addCriterion("`status` =", value, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusNotEqualTo(Integer value) {
+            addCriterion("`status` <>", value, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusGreaterThan(Integer value) {
+            addCriterion("`status` >", value, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusGreaterThanOrEqualTo(Integer value) {
+            addCriterion("`status` >=", value, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusLessThan(Integer value) {
+            addCriterion("`status` <", value, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusLessThanOrEqualTo(Integer value) {
+            addCriterion("`status` <=", value, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusIn(List<Integer> values) {
+            addCriterion("`status` in", values, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusNotIn(List<Integer> values) {
+            addCriterion("`status` not in", values, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusBetween(Integer value1, Integer value2) {
+            addCriterion("`status` between", value1, value2, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andStatusNotBetween(Integer value1, Integer value2) {
+            addCriterion("`status` not between", value1, value2, "status");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdIsNull() {
+            addCriterion("`dirver_id` is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdIsNotNull() {
+            addCriterion("`dirver_id` is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdEqualTo(Long value) {
+            addCriterion("`dirver_id` =", value, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdNotEqualTo(Long value) {
+            addCriterion("`dirver_id` <>", value, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdGreaterThan(Long value) {
+            addCriterion("`dirver_id` >", value, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdGreaterThanOrEqualTo(Long value) {
+            addCriterion("`dirver_id` >=", value, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdLessThan(Long value) {
+            addCriterion("`dirver_id` <", value, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdLessThanOrEqualTo(Long value) {
+            addCriterion("`dirver_id` <=", value, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdIn(List<Long> values) {
+            addCriterion("`dirver_id` in", values, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdNotIn(List<Long> values) {
+            addCriterion("`dirver_id` not in", values, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdBetween(Long value1, Long value2) {
+            addCriterion("`dirver_id` between", value1, value2, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDirverIdNotBetween(Long value1, Long value2) {
+            addCriterion("`dirver_id` not between", value1, value2, "dirverId");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedIsNull() {
+            addCriterion("`deleted` is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedIsNotNull() {
+            addCriterion("`deleted` is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedEqualTo(Boolean value) {
+            addCriterion("`deleted` =", value, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedNotEqualTo(Boolean value) {
+            addCriterion("`deleted` <>", value, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedGreaterThan(Boolean value) {
+            addCriterion("`deleted` >", value, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedGreaterThanOrEqualTo(Boolean value) {
+            addCriterion("`deleted` >=", value, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedLessThan(Boolean value) {
+            addCriterion("`deleted` <", value, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedLessThanOrEqualTo(Boolean value) {
+            addCriterion("`deleted` <=", value, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedIn(List<Boolean> values) {
+            addCriterion("`deleted` in", values, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedNotIn(List<Boolean> values) {
+            addCriterion("`deleted` not in", values, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedBetween(Boolean value1, Boolean value2) {
+            addCriterion("`deleted` between", value1, value2, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andDeletedNotBetween(Boolean value1, Boolean value2) {
+            addCriterion("`deleted` not between", value1, value2, "deleted");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedIsNull() {
+            addCriterion("`created` is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedIsNotNull() {
+            addCriterion("`created` is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedEqualTo(Date value) {
+            addCriterion("`created` =", value, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedNotEqualTo(Date value) {
+            addCriterion("`created` <>", value, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedGreaterThan(Date value) {
+            addCriterion("`created` >", value, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedGreaterThanOrEqualTo(Date value) {
+            addCriterion("`created` >=", value, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedLessThan(Date value) {
+            addCriterion("`created` <", value, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedLessThanOrEqualTo(Date value) {
+            addCriterion("`created` <=", value, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedIn(List<Date> values) {
+            addCriterion("`created` in", values, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedNotIn(List<Date> values) {
+            addCriterion("`created` not in", values, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedBetween(Date value1, Date value2) {
+            addCriterion("`created` between", value1, value2, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andCreatedNotBetween(Date value1, Date value2) {
+            addCriterion("`created` not between", value1, value2, "created");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedIsNull() {
+            addCriterion("`updated` is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedIsNotNull() {
+            addCriterion("`updated` is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedEqualTo(Date value) {
+            addCriterion("`updated` =", value, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedNotEqualTo(Date value) {
+            addCriterion("`updated` <>", value, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedGreaterThan(Date value) {
+            addCriterion("`updated` >", value, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedGreaterThanOrEqualTo(Date value) {
+            addCriterion("`updated` >=", value, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedLessThan(Date value) {
+            addCriterion("`updated` <", value, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedLessThanOrEqualTo(Date value) {
+            addCriterion("`updated` <=", value, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedIn(List<Date> values) {
+            addCriterion("`updated` in", values, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedNotIn(List<Date> values) {
+            addCriterion("`updated` not in", values, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedBetween(Date value1, Date value2) {
+            addCriterion("`updated` between", value1, value2, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andUpdatedNotBetween(Date value1, Date value2) {
+            addCriterion("`updated` not between", value1, value2, "updated");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionIsNull() {
+            addCriterion("`version` is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionIsNotNull() {
+            addCriterion("`version` is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionEqualTo(Integer value) {
+            addCriterion("`version` =", value, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionNotEqualTo(Integer value) {
+            addCriterion("`version` <>", value, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionGreaterThan(Integer value) {
+            addCriterion("`version` >", value, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionGreaterThanOrEqualTo(Integer value) {
+            addCriterion("`version` >=", value, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionLessThan(Integer value) {
+            addCriterion("`version` <", value, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionLessThanOrEqualTo(Integer value) {
+            addCriterion("`version` <=", value, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionIn(List<Integer> values) {
+            addCriterion("`version` in", values, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionNotIn(List<Integer> values) {
+            addCriterion("`version` not in", values, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionBetween(Integer value1, Integer value2) {
+            addCriterion("`version` between", value1, value2, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andVersionNotBetween(Integer value1, Integer value2) {
+            addCriterion("`version` not between", value1, value2, "version");
+            return (Criteria) this;
+        }
+
+        public Criteria andPassengerLikeInsensitive(String value) {
+            addCriterion("upper(`passenger`) like", value.toUpperCase(), "passenger");
+            return (Criteria) this;
+        }
+    }
+
+    /**
+     * This class was generated by MyBatis Generator.
+     * This class corresponds to the database table demand_order
+     *
+     * @mbg.generated do_not_delete_during_merge
+     */
+    public static class Criteria extends GeneratedCriteria {
+
+        protected Criteria() {
+            super();
+        }
+    }
+
+    /**
+     * This class was generated by MyBatis Generator.
+     * This class corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    public static class Criterion {
+        private String condition;
+
+        private Object value;
+
+        private Object secondValue;
+
+        private boolean noValue;
+
+        private boolean singleValue;
+
+        private boolean betweenValue;
+
+        private boolean listValue;
+
+        private String typeHandler;
+
+        public String getCondition() {
+            return condition;
+        }
+
+        public Object getValue() {
+            return value;
+        }
+
+        public Object getSecondValue() {
+            return secondValue;
+        }
+
+        public boolean isNoValue() {
+            return noValue;
+        }
+
+        public boolean isSingleValue() {
+            return singleValue;
+        }
+
+        public boolean isBetweenValue() {
+            return betweenValue;
+        }
+
+        public boolean isListValue() {
+            return listValue;
+        }
+
+        public String getTypeHandler() {
+            return typeHandler;
+        }
+
+        protected Criterion(String condition) {
+            super();
+            this.condition = condition;
+            this.typeHandler = null;
+            this.noValue = true;
+        }
+
+        protected Criterion(String condition, Object value, String typeHandler) {
+            super();
+            this.condition = condition;
+            this.value = value;
+            this.typeHandler = typeHandler;
+            if (value instanceof List<?>) {
+                this.listValue = true;
+            } else {
+                this.singleValue = true;
+            }
+        }
+
+        protected Criterion(String condition, Object value) {
+            this(condition, value, null);
+        }
+
+        protected Criterion(String condition, Object value, Object secondValue, String typeHandler) {
+            super();
+            this.condition = condition;
+            this.value = value;
+            this.secondValue = secondValue;
+            this.typeHandler = typeHandler;
+            this.betweenValue = true;
+        }
+
+        protected Criterion(String condition, Object value, Object secondValue) {
+            this(condition, value, secondValue, null);
+        }
+    }
+}
