+++ db-practice/src/main/java/com/si/practice/dbpractice/infrature/mysql/mapper/DemandOrderMapper.java
@@ -1,0 +1,98 @@
+package com.si.practice.dbpractice.infrature.mysql.mapper;
+
+
+import com.si.practice.dbpractice.infrature.mysql.entity.DemandOrder;
+import com.si.practice.dbpractice.infrature.mysql.entity.DemandOrderExample;
+import org.apache.ibatis.annotations.Param;
+
+import java.util.List;
+
+public interface DemandOrderMapper {
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    long countByExample(DemandOrderExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    int deleteByExample(DemandOrderExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    int deleteByPrimaryKey(Long id);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    int insert(DemandOrder record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    int insertSelective(DemandOrder record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    List<DemandOrder> selectByExample(DemandOrderExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    DemandOrder selectByPrimaryKey(Long id);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    int updateByExampleSelective(@Param("record") DemandOrder record, @Param("example") DemandOrderExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    int updateByExample(@Param("record") DemandOrder record, @Param("example") DemandOrderExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    int updateByPrimaryKeySelective(DemandOrder record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table demand_order
+     *
+     * @mbg.generated
+     */
+    int updateByPrimaryKey(DemandOrder record);
+}
