+++ db-practice/src/main/java/com/si/practice/dbpractice/infrature/mysql/t.java
@@ -1,0 +1,4 @@
+package practice.infrature.mysql;
+
+public class t {
+}
