+++ db-practice/src/main/resources/application.properties
@@ -1,0 +1,10 @@
+server.port=9999
+
+#mybatis.typeAliasesPackage=com.si.OrderEntity
+
+mybatis.mapperLocations=classpath:mapper/*/*.xml,classpath:mapper/*.xml
+
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.url=jdbc:mysql://127.0.0.1:3306/tf-manager?useUnicode=true&characterEncoding=utf-8&serverTimezone=GMT%2B8&useSSL=false&autoReconnect=true
+spring.datasource.username=root
+spring.datasource.password=root123456
