+++ db-practice/src/main/resources/script/DBInitScript.sql
