+++ db-practice/src/main/resources/script/my.cnf
