+++ db-practice/src/test/java/com/si/practice/dbpractice/DBPracticeApplicationTests.java
@@ -1,0 +1,13 @@
+package com.si.practice.dbpractice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DBPracticeApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
