--- src/main/java/com/si/practice/app/t.java
@@ -1,4 +1,0 @@
-package com.si.practice.app;
-
-public class t {
-}
