--- src/main/java/com/si/practice/client/order/api/OrderCreateService.java
@@ -1,12 +1,0 @@
-package com.si.practice.client.order.api;
-
-import com.si.practice.common.order.OrderDTO;
-
-import java.util.List;
-
-public interface OrderCreateService {
-
-    Long createOrder(OrderDTO orderDTO);
-
-    Boolean batchInsert(List<OrderDTO> orderDTOS);
-}
