--- src/main/java/com/si/practice/client/order/api/OrderUpdateService.java
@@ -1,6 +1,0 @@
-package com.si.practice.client.order.api;
-
-public interface OrderUpdateService {
-
-    Boolean updateOrderById(Long orderId, Integer status);
-}
