--- src/main/java/com/si/practice/client/order/request/QueryOrderRequest.java
@@ -1,6 +1,0 @@
-package com.si.practice.client.order.request;
-
-import java.io.Serializable;
-
-public class QueryOrderRequest implements Serializable {
-}
