--- src/main/java/com/si/practice/common/order/OrderDTO.java
@@ -1,34 +1,0 @@
-package com.si.practice.common.order;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-import java.util.Date;
-
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class OrderDTO implements Serializable {
-
-
-    private Long id;
-
-    private String passenger;
-
-    private Long prodId;
-
-    private Integer status;
-
-    private Long driverId;
-
-    private Integer version;
-
-    private Boolean deleted;
-
-    private Date created;
-
-    private Date updated;
-}
