--- src/main/java/com/si/practice/infrature/mysql/mapper/DemandOrderMapper.java
@@ -1,96 +1,0 @@
-package com.si.practice.infrature.mysql.mapper;
-
-import com.si.practice.infrature.mysql.entity.DemandOrder;
-import com.si.practice.infrature.mysql.entity.DemandOrderExample;
-import java.util.List;
-import org.apache.ibatis.annotations.Param;
-
-public interface DemandOrderMapper {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    long countByExample(DemandOrderExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    int deleteByExample(DemandOrderExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    int deleteByPrimaryKey(Long id);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    int insert(DemandOrder record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    int insertSelective(DemandOrder record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    List<DemandOrder> selectByExample(DemandOrderExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    DemandOrder selectByPrimaryKey(Long id);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    int updateByExampleSelective(@Param("record") DemandOrder record, @Param("example") DemandOrderExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    int updateByExample(@Param("record") DemandOrder record, @Param("example") DemandOrderExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    int updateByPrimaryKeySelective(DemandOrder record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table demand_order
-     *
-     * @mbg.generated
-     */
-    int updateByPrimaryKey(DemandOrder record);
-}
