--- src/main/java/com/si/practice/infrature/mysql/t.java
@@ -1,4 +1,0 @@
-package com.si.practice.infrature.mysql;
-
-public class t {
-}
