--- src/main/resources/application.properties
@@ -1,10 +1,0 @@
-server.port=9999
-
-#mybatis.typeAliasesPackage=com.si.practice.infrature.mysql.entity.OrderEntity
-
-mybatis.mapperLocations=classpath:mappers/*/*.xml,classpath:mappers/*.xml
-
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://127.0.0.1:3306/tf-manager?useUnicode=true&characterEncoding=utf-8&serverTimezone=GMT%2B8&useSSL=false&autoReconnect=true
-spring.datasource.username=root
-spring.datasource.password=root123456
