--- src/main/resources/mappers/order/OrderBaseMapper.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.si.practice.infrature.mysql.mapper.OrderBaseMapper">
-  <resultMap id="BaseResultMap" type="com.si.practice.infrature.mysql.entity.OrderEntity">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-    -->
-    <id column="id" jdbcType="BIGINT" property="id" />
-    <result column="passenger" jdbcType="VARCHAR" property="passenger" />
-    <result column="prod_id" jdbcType="BIGINT" property="prodId" />
-    <result column="status" jdbcType="INTEGER" property="status" />
-    <result column="driver_id" jdbcType="BIGINT" property="driverId" />
-    <result column="deleted" jdbcType="BIT" property="deleted" />
-    <result column="created" jdbcType="TIMESTAMP" property="created" />
-    <result column="updated" jdbcType="TIMESTAMP" property="updated" />
-    <result column="version" jdbcType="INTEGER" property="version" />
-  </resultMap>
-
-  <sql id="Base_Column_List">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-    -->
-    `id`, `passenger`, `prod_id`, `status`, `driver_id`, `deleted`, `created`, `updated`, 
-    `version`
-  </sql>
-    <insert id="batchInsert" parameterType="java.util.List">
-
-        insert ignore into demand_order (`passenger`, `prod_id`, `status`,
-        `driver_id`, `deleted`, `created`,
-        `updated`, `version`)
-        values
-        <foreach collection="orders" item="order" separator=",">
-            <trim prefix="(" suffix=")" suffixOverrides=",">
-                #{order.passenger,jdbcType=VARCHAR},
-                #{order.prodId,jdbcType=BIGINT},
-                #{order.status,jdbcType=INTEGER},
-                #{order.driverId,jdbcType=BIGINT},
-                #{order.deleted,jdbcType=BIT},
-                #{order.created,jdbcType=TIMESTAMP},
-                #{order.updated,jdbcType=TIMESTAMP},
-                #{order.version,jdbcType=INTEGER}
-            </trim>
-        </foreach>
-
-    </insert>
-</mapper>
