--- src/test/java/com/si/practice/PracticeApplicationTests.java
@@ -1,13 +1,0 @@
-package com.si.practice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PracticeApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
